verso com a tela de carrinho, checkout e mapa atualizada, com a taxa de entrega adicionada

+++ Xpress/Model/FacturaActual.swift
@@ -23,7 +23,7 @@ class FacturaActual: Codable {
       var dataPedido: String = ""
       var itens = [Iten]()
       var total: Double = 0.0
-      var taxaboy: Int = 0
+    var taxaboy: Double = 0.0
      
       init(identificacaoPagamento: String, entidade: String, idFactura: Int, metododPagamento: String, horaEntregueMotoboy: String, horaRecebidoCliente: String, estado: String, estadoPagamento: String, clienteID: Int, dataPagamento: String, dataPedido: String, itens: [Iten], total: Double, taxaboy: Int) {
